reproducible regression tests
