+++ Cargo.toml
@@ -24,4 +24,5 @@ tokio = "0.1.8"
 default = ["std"]
 std = ["parity-scale-codec/std", "num/std", "parking_lot"]
 derive-codec = ["parity-scale-codec"]
-test-helpers = []
+test-helpers = ["fuzz-helpers"]
+fuzz-helpers = []
